# Conflicts:
#	db/xmlconv-struct.xml
#	src/main/resources/quartz.properties

+++ src/main/java/eionet/gdem/web/listeners/WebSessionListener.java
@@ -23,8 +23,9 @@ public class WebSessionListener implements HttpSessionListener {
     @Override
     public void sessionDestroyed(HttpSessionEvent httpSessionEvent) {
         String id = httpSessionEvent.getSession().getId();
-        String sessionDir = Properties.appRootFolder + File.separator + "tmpfile" + File.separator + id;
+        String sessionDir = Properties.appRootFolder + "/tmpfile/" + id;
         File dir = new File(sessionDir);
+        if ( !dir.exists()) return;
         for(File file: dir.listFiles()) {
             file.delete();
         }
